requirements.txt
fastapi[all]
uvicorn
jinja2
google-generativeai
kiteconnect
